refactor(transform): split oxc template dependency collection
