+++ PLAN.tasks/TASK-04-split-template-dependencies.md
@@ -1,6 +1,6 @@
 # TASK-04: Split collectOxcTemplateDependencies
 
-Status: pending
+Status: done
 
 ## Goal
 
@@ -13,19 +13,22 @@ Split `collectOxcTemplateDependencies.ts` into focused modules while keeping its
 
 ## Checklist
 
-- [ ] Identify public exports and internal responsibilities.
-- [ ] Extract bindings/scope analysis.
-- [ ] Extract evaluator logic.
-- [ ] Extract expression dependency collection.
-- [ ] Extract static binding helpers.
-- [ ] Leave the original file as a small facade.
-- [ ] Confirm production `.ts` files are below 1000 lines where this task touches them.
-- [ ] Run focused tests.
-- [ ] Update `PLAN.md` status and progress log.
+- [x] Identify public exports and internal responsibilities.
+- [x] Extract bindings/scope analysis.
+- [x] Extract evaluator logic.
+- [x] Extract expression dependency collection.
+- [x] Extract static binding helpers.
+- [x] Leave the original file as a small facade.
+- [x] Confirm production `.ts` files are below 1000 lines where this task touches them.
+- [x] Run focused tests.
+- [x] Update `PLAN.md` status and progress log.
 
 ## Progress Log
 
 - 2026-05-08 14:36 EEST: Task created by scaffold.
+- 2026-05-08 15:17 EEST: Started task after committing shared Oxc utilities. Inspecting top-level structure and dependency direction before moving code.
+- 2026-05-08 15:37 EEST: Split the module into `types`, `staticBindings`, `scopeAnalysis`, `staticEvaluator`, `expressionReplacements`, and `expressionExtraction`. Kept `collectOxcTemplateDependencies.ts` as a facade.
+- 2026-05-08 15:43 EEST: Verified split with focused tests, transform type build, transform lint, size guard, and script lint. Removed `collectOxcTemplateDependencies.ts` from the legacy size allowlist because the facade and new modules are under 1000 lines.
 
 ## Context Recovery Notes
 
@@ -34,6 +37,10 @@ Keep call sites unchanged unless a local import path update is required. This ta
 ## Test Commands
 
 - `cd wyw-in-js/packages/transform && bun test src/__tests__/collectOxcTemplateDependencies.test.ts src/__tests__/transform.static-import-values.test.ts`
+- `cd wyw-in-js && bun run --filter @wyw-in-js/transform build:types`
+- `cd wyw-in-js && bun run --filter @wyw-in-js/transform lint`
+- `cd wyw-in-js && bun run check:ts-size`
+- `cd wyw-in-js && bun run lint:scripts`
 
 ## Done Criteria
