+++ scripts/check-production-ts-size.mjs
@@ -18,7 +18,6 @@ const LEGACY_OVERSIZED_FILES = new Map([
   ],
   ['packages/transform/src/utils/applyOxcProcessors.ts', 2937],
   ['packages/transform/src/utils/collectOxcExportsAndImports.ts', 1436],
-  ['packages/transform/src/utils/collectOxcTemplateDependencies.ts', 2802],
   ['packages/transform/src/utils/oxcPreevalTransforms.ts', 1794],
   ['packages/transform/src/utils/oxcShaker.ts', 1082],
   ['packages/vite/src/index.ts', 1210],
